+++ Cargo.toml
@@ -35,6 +35,7 @@ members = [
     "core-primitives/node-api/metadata-provider",
     "core-primitives/nonce-cache",
     "core-primitives/ocall-api",
+    "core-primitives/pallet-storage",
     "core-primitives/primitives-cache",
     "core-primitives/rpc",
     "core-primitives/settings",
@@ -50,7 +51,6 @@ members = [
     "core-primitives/substrate-sgx/environmental",
     "core-primitives/substrate-sgx/externalities",
     "core-primitives/substrate-sgx/sp-io",
-    "core-primitives/teerex-storage",
     "core-primitives/test",
     "core-primitives/time-utils",
     "core-primitives/top-pool",
