--- core-primitives/enclave-bridge-storage/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "itp-enclave-bridge-storage"
-version = "0.9.0"
-authors = ["Integritee AG <hello@integritee.network>"]
-edition = "2021"
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
-sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
-
-#local deps
-itp-storage = { path = "../storage", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "codec/std",
-    "sp-std/std",
-    "itp-storage/std",
-]
